pyproject.toml
[tool.poetry]
name = "bluedawn"
version = "0.7.1"
description = "a discord bot"
authors = ["Kent D <lolexplode@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.3"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
black = "^21.6b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
